backend/requirements.txt
django<3.3.0
djangorestframework<3.13.0
markdown<3.4.0
django-filter<2.5.0
django-redis<5.1.0
celery<5.2.0
psycopg2-binary